+++ src/Database/Migrations/2024_05_08_043115_create_postal_codes_table.php
@@ -1,0 +1,27 @@
+<?php
+
+use Illuminate\Database\Migrations\Migration;
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Support\Facades\Schema;
+
+return new class extends Migration
+{
+    public function up(): void
+    {
+        Schema::create('postal_codes', function (Blueprint $table) {
+            $table->id();
+            $table->integer('PostalCodeID');
+            $table->string('PostalCode');
+            $table->string('City');
+            $table->string('State');
+            $table->integer('CountryID');
+            $table->string('County');
+            $table->timestamps();
+        });
+    }
+
+    public function down(): void
+    {
+        Schema::dropIfExists('postal_codes');
+    }
+};